Unit tests were moved to the relevant AdViewUtilsTests class
